[tgm][api][break] Major api structure update with some breaking changes.
Moved a few methods, added a few missing documentation pieces.
