+++ src/main/java/org/frikadelki/ash/telegram/api/chat/TgmBotApiChat.java
@@ -21,8 +21,18 @@ public interface TgmBotApiChat {
 		public static String PARSE_MODE_MARKDOWN = "Markdown";
 		public static String PARSE_MODE_HTML = "HTML";
 
+		/**
+		 * Required
+		 * Unique identifier for the target chat or username of the
+		 * target channel (in the format @channelusername)
+		 */
 		@SerializedName("chat_id")
 		@NonNull private final Long chatId;
+
+		/**
+		 * Required
+		 * Text of the message to be sent
+		 */
 		@NonNull private final String text;
 
 		/**
